reshape(), append_rows(), and unseq funcs
added nodiscard to a bunch of stuff, including template expr functions.
added noexcept to more functions that could use it.

added reshape() - equivalent to resize_flat() but requires no change in total number of elements.
added append_rows() - same as append_cols() but for rows, and therefore way simpler.

fixed unseq logic for vectorized functions.
previously, all vectorized functions were assumed to be sequenced as a conservative assumption.
now a template param is passed to vectorize() to determine if it is unseq or not.
additionally, due to the name 'vectorize()' this defaults to unseq if not provided.